github-actions: Don't re-run on tags
Support running github actions on any branch but do
not run if the tag changes or is added to a specific
commit.  This prevents extra actions from being run
when a branch is created from an existing commit and
only results in a new actions run if/when a new commit
is pushed on the resulting branch.